requirements.txt
openai~=1.35.10
python-dotenv~=1.0.1
azure-search-documents==11.6.0b1
azure-identity
azure-storage-blob
fastapi
uvicorn
jinja2
python-multipart
pyarrow
pandas~=2.2.2
langchain~=0.2.8
langchain-openai
psycopg2-binary
huggingface_hub
peft
transformers
msal
requests
fitz
pymupdf
nltk~=3.8.1
llama-index-embeddings-openai
langchain-text-splitters
langchain_experimental
tiktoken~=0.7.0
numpy~=1.26.4
tqdm~=4.66.4
matplotlib
seaborn
scikit-learn
scipy
